fix(utility): make typing more convenient in text and font relative code

+++ libraries/utility/headers/utility/graphic/text/glyph.hpp
@@ -42,7 +42,7 @@ namespace utility::graphic
 		 * @brief Advance value, specifying the horizontal distance to advance
 		 * the cursor after rendering the glyph.
 		 */
-		uint32_t advance;
+		float advance;
 
 		/**
 		 * @brief Minimum UV coordinates for the glyph in the font atlas.
